refactor: final main.ts cleanup — extract remaining inline handlers
Move session register/unregister to session-lifecycle.ts, skill record
delete with disk cleanup to skills.ts factory, team utility routes
(backfill/retry/upgrade-worker) to team-connect.ts, and backup config
routes to backup.ts. Remove dead imports (zod, team-sync-client,
outbox queries, session queries, notify, cleanStaleBuffers).
